reduce number of examples and overlap between them
