+++ examples/uma_ase_recipes.py
@@ -9,10 +9,11 @@
 # ---
 
 # %% [markdown]
-# # ASE Recipes with UMA-JAX
+# # Advanced ASE Recipes with Pretrained UMA
 #
-# A collection of common atomistic simulation recipes using the UMA
-# calculator with ASE. Each recipe is self-contained.
+# Advanced atomistic simulation recipes using a pretrained UMA model
+# with ASE. For basic getting-started workflows (relaxation, EOS,
+# molecule optimization, trajectory I/O), see `uma_pretrained_ase.py`.
 #
 # **Prerequisites:**
 # ```bash
@@ -23,130 +24,7 @@
 import numpy as np
 
 # %% [markdown]
-# ## Recipe 1: Single-Point Calculation
-
-# %%
-def recipe_single_point():
-  """Compute energy and forces for a given structure."""
-  from ase.build import bulk
-  from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
-
-  atoms = bulk('Cu', 'fcc', a=3.615, cubic=True)
-
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-  atoms.calc = UMACalculator(config=cfg, task_name='omat')
-
-  energy = atoms.get_potential_energy()
-  forces = atoms.get_forces()
-  max_force = np.max(np.abs(forces))
-
-  print(f"System: {atoms.get_chemical_formula()}")
-  print(f"Energy: {energy:.6f} eV")
-  print(f"Max force: {max_force:.6f} eV/A")
-  print(f"Forces shape: {forces.shape}")
-  return atoms
-
-try:
-  atoms = recipe_single_point()
-except ImportError as e:
-  print(f"Skipping: {e}")
-
-# %% [markdown]
-# ## Recipe 2: Geometry Optimization (BFGS)
-
-# %%
-def recipe_bfgs_relaxation():
-  """Relax atomic positions with BFGS."""
-  from ase.build import molecule
-  from ase.optimize import BFGS
-  from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
-
-  # Water molecule in a box
-  atoms = molecule('H2O')
-  atoms.center(vacuum=5.0)
-  atoms.pbc = False
-
-  # Perturb positions
-  rng = np.random.default_rng(42)
-  atoms.positions += rng.normal(scale=0.1, size=atoms.positions.shape)
-
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=6.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-  atoms.calc = UMACalculator(config=cfg, task_name='omol')
-
-  print(f"Initial energy: {atoms.get_potential_energy():.6f} eV")
-
-  opt = BFGS(atoms, logfile=None)
-  opt.run(fmax=0.05, steps=30)
-
-  print(f"Final energy:   {atoms.get_potential_energy():.6f} eV")
-  print(f"Steps taken:    {opt.nsteps}")
-  print(f"Max force:      {np.max(np.abs(atoms.get_forces())):.4f} eV/A")
-
-try:
-  recipe_bfgs_relaxation()
-except ImportError as e:
-  print(f"Skipping: {e}")
-
-# %% [markdown]
-# ## Recipe 3: Equation of State (EOS) Curve
-
-# %%
-def recipe_eos():
-  """Compute energy vs volume curve for equation of state fitting."""
-  from ase.build import bulk
-  from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
-
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-
-  # Scan lattice constants around equilibrium
-  a0 = 3.615  # Cu FCC equilibrium
-  strains = np.linspace(-0.05, 0.05, 11)
-  volumes = []
-  energies = []
-
-  for strain in strains:
-    a = a0 * (1 + strain)
-    atoms = bulk('Cu', 'fcc', a=a)
-    atoms.calc = UMACalculator(config=cfg, task_name='omat')
-
-    E = atoms.get_potential_energy()
-    V = atoms.get_volume()
-
-    volumes.append(V)
-    energies.append(E)
-    print(f"  a={a:.3f} A, V={V:.2f} A^3, E={E:.6f} eV")
-
-  # Fit Birch-Murnaghan EOS
-  try:
-    from ase.eos import EquationOfState
-    eos = EquationOfState(volumes, energies, eos='birchmurnaghan')
-    v0, e0, B = eos.fit()
-    print(f"\nEOS fit: V0={v0:.2f} A^3, E0={e0:.6f} eV, B={B/1.602e-19*1e30/1e9:.1f} GPa")
-  except Exception as e:
-    print(f"\nEOS fit skipped: {e}")
-
-try:
-  recipe_eos()
-except ImportError as e:
-  print(f"Skipping: {e}")
-
-# %% [markdown]
-# ## Recipe 4: Surface Slab + Adsorbate
+# ## Recipe 1: Surface Slab + Adsorbate
 
 # %%
 def recipe_surface_relaxation():
@@ -155,25 +33,16 @@ def recipe_surface_relaxation():
   from ase.constraints import FixAtoms
   from ase.optimize import BFGS
   from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
 
-  # Build 3-layer Cu(111) slab with vacuum
   slab = fcc111('Cu', size=(2, 2, 3), vacuum=10.0)
 
-  # Fix bottom layer
   z_positions = slab.positions[:, 2]
   bottom_layer = z_positions < z_positions.min() + 1.0
   slab.set_constraint(FixAtoms(mask=bottom_layer))
 
-  # Add CO adsorbate on top
   add_adsorbate(slab, 'C', height=1.8, position='ontop')
 
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-  slab.calc = UMACalculator(config=cfg, task_name='oc20')
+  slab.calc = UMACalculator(checkpoint_path='uma-s-1p1', task_name='oc20')
 
   print(f"Slab: {slab.get_chemical_formula()}, {len(slab)} atoms")
   print(f"Fixed atoms: {sum(bottom_layer)}")
@@ -191,7 +60,7 @@ except ImportError as e:
   print(f"Skipping: {e}")
 
 # %% [markdown]
-# ## Recipe 5: Molecular Dynamics with ASE
+# ## Recipe 2: Molecular Dynamics with ASE
 
 # %%
 def recipe_ase_md():
@@ -200,18 +69,10 @@ def recipe_ase_md():
   from ase.md.langevin import Langevin
   from ase import units
   from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
 
   atoms = bulk('Cu', 'fcc', a=3.615, cubic=True).repeat((2, 2, 2))
+  atoms.calc = UMACalculator(checkpoint_path='uma-s-1p1', task_name='omat')
 
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-  atoms.calc = UMACalculator(config=cfg, task_name='omat')
-
-  # Langevin NVT at 300K
   dyn = Langevin(
     atoms,
     timestep=1.0 * units.fs,
@@ -239,28 +100,20 @@ except ImportError as e:
   print(f"Skipping: {e}")
 
 # %% [markdown]
-# ## Recipe 6: Elastic Constants
+# ## Recipe 3: Elastic Constants
 
 # %%
 def recipe_elastic():
   """Compute elastic constants via strain-stress relation."""
   from ase.build import bulk
   from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
 
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
+  calc = UMACalculator(checkpoint_path='uma-s-1p1', task_name='omat')
 
-  # Reference structure
   atoms = bulk('Cu', 'fcc', a=3.615, cubic=True)
-  atoms.calc = UMACalculator(config=cfg, task_name='omat')
+  atoms.calc = calc
   E0 = atoms.get_potential_energy()
-  V0 = atoms.get_volume()
 
-  # Apply small strains and compute energy
   delta = 0.005
   strains = [-2*delta, -delta, 0, delta, 2*delta]
 
@@ -269,7 +122,7 @@ def recipe_elastic():
     strained = atoms.copy()
     cell = strained.get_cell()
     strained.set_cell(cell * (1 + eps), scale_atoms=True)
-    strained.calc = UMACalculator(config=cfg, task_name='omat')
+    strained.calc = calc
     E = strained.get_potential_energy()
     print(f"  eps={eps:+.4f}: E={E:.6f} eV (dE={E-E0:+.6f})")
 
@@ -279,7 +132,7 @@ except ImportError as e:
   print(f"Skipping: {e}")
 
 # %% [markdown]
-# ## Recipe 7: Batch Screening
+# ## Recipe 4: Batch Screening
 #
 # Evaluate multiple structures efficiently by reusing the calculator.
 
@@ -288,16 +141,9 @@ def recipe_batch_screening():
   """Screen multiple structures for their energies."""
   from ase.build import bulk
   from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
 
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-  calc = UMACalculator(config=cfg, task_name='omat')
+  calc = UMACalculator(checkpoint_path='uma-s-1p1', task_name='omat')
 
-  # Screen FCC metals at experimental lattice constants
   metals = {
     'Cu': 3.615,
     'Ag': 4.085,
@@ -319,7 +165,6 @@ def recipe_batch_screening():
     results[element] = e_per_atom
     print(f"{element:>5} {a:>8.3f} {e_per_atom:>12.6f}")
 
-  # Rank by energy
   sorted_metals = sorted(results.items(), key=lambda x: x[1])
   print(f"\nMost stable: {sorted_metals[0][0]} ({sorted_metals[0][1]:.4f} eV/atom)")
 
@@ -329,22 +174,13 @@ except ImportError as e:
   print(f"Skipping: {e}")
 
 # %% [markdown]
-# ## Recipe 8: Charged/Spin Systems (omol)
+# ## Recipe 5: Charged/Spin Systems (omol)
 
 # %%
 def recipe_charged_system():
   """Handle charged molecules and spin states."""
   from ase import Atoms
   from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
-
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=6.0, edge_channels=32,
-    num_distance_basis=64,
-    dataset_list=['oc20', 'omol', 'omat', 'odac', 'omc'],
-    use_dataset_embedding=True,
-  )
 
   # O2 molecule: neutral triplet
   o2 = Atoms('O2', positions=[[0, 0, 0], [0, 0, 1.21]])
@@ -352,7 +188,7 @@ def recipe_charged_system():
   o2.info['charge'] = 0
   o2.info['spin'] = 3  # triplet: spin multiplicity = 2S+1 = 3
 
-  o2.calc = UMACalculator(config=cfg, task_name='omol')
+  o2.calc = UMACalculator(checkpoint_path='uma-s-1p1', task_name='omol')
   E_neutral = o2.get_potential_energy()
   print(f"O2 (neutral, triplet): E = {E_neutral:.4f} eV")
 
@@ -360,7 +196,7 @@ def recipe_charged_system():
   o2_anion = o2.copy()
   o2_anion.info['charge'] = -1
   o2_anion.info['spin'] = 2  # doublet
-  o2_anion.calc = UMACalculator(config=cfg, task_name='omol')
+  o2_anion.calc = UMACalculator(checkpoint_path='uma-s-1p1', task_name='omol')
   E_anion = o2_anion.get_potential_energy()
   print(f"O2- (anion, doublet):  E = {E_anion:.4f} eV")
   print(f"Electron affinity:     {E_neutral - E_anion:.4f} eV")
@@ -369,65 +205,3 @@ try:
   recipe_charged_system()
 except ImportError as e:
   print(f"Skipping: {e}")
-
-# %% [markdown]
-# ## Recipe 9: Trajectory I/O
-
-# %%
-def recipe_trajectory():
-  """Save and load optimization trajectories."""
-  from ase.build import bulk
-  from ase.optimize import BFGS
-  from ase.io import read, write
-  from ase.io.trajectory import Trajectory
-  from jax_md._nn.uma.ase_calculator import UMACalculator
-  from jax_md._nn.uma.model import UMAConfig
-
-  cfg = UMAConfig(
-    sphere_channels=32, lmax=2, mmax=2, num_layers=1,
-    hidden_channels=32, cutoff=5.0, edge_channels=32,
-    num_distance_basis=64, use_dataset_embedding=False,
-  )
-
-  atoms = bulk('Si', 'diamond', a=5.43)
-  rng = np.random.default_rng(42)
-  atoms.positions += rng.normal(scale=0.05, size=atoms.positions.shape)
-
-  atoms.calc = UMACalculator(config=cfg, task_name='omat')
-
-  # Save trajectory
-  traj_file = '/tmp/uma_relax.traj'
-  with Trajectory(traj_file, 'w', atoms) as traj:
-    opt = BFGS(atoms, logfile=None)
-    opt.attach(traj.write, interval=1)
-    opt.run(fmax=0.1, steps=10)
-
-  # Read trajectory
-  frames = read(traj_file, index=':')
-  print(f"Trajectory: {len(frames)} frames saved to {traj_file}")
-  for i, frame in enumerate(frames):
-    print(f"  Frame {i}: E={frame.get_potential_energy():.4f} eV")
-
-  # Export to XYZ for visualization
-  xyz_file = '/tmp/uma_relax.xyz'
-  write(xyz_file, frames)
-  print(f"XYZ written to {xyz_file}")
-
-try:
-  recipe_trajectory()
-except ImportError as e:
-  print(f"Skipping: {e}")
-
-# %% [markdown]
-# ## Using Pretrained Models
-#
-# All recipes above use random weights for demonstration. To use
-# pretrained UMA weights from FairChem:
-#
-# ```python
-# calc = UMACalculator(
-#     checkpoint_path='path/to/uma_sm_conserve.pt',
-#     task_name='omat',  # or 'omol', 'oc20', etc.
-# )
-# atoms.calc = calc
-# ```
